feat: support changing Issue title and description
