+++ README.md
@@ -96,6 +96,17 @@ renovate-issue-action.yaml
 
 ```yaml
 renovate_login: 'renovate[bot]'
+issue:
+  title: 'Renovate Automerge Failure({{.RepoOwner}}/{{.RepoName}}): {{if .Metadata.GroupName}}{{.Metadata.GroupName}}{{else}}{{.Metadata.PackageName}}{{.Metadata.DepName}}{{end}} {{if .Metadata.PackageFileDir}}({{.Metadata.PackageFileDir}}){{end}}'
+  description_header: |
+    _This pull request was created by [renovate-issue-action](https://github.com/suzuki-shunsuke/renovate-issue-action)._
+
+    :warning: Please don't edit the Issue title, because renovate-issue-action searches issue with Issue title.
+
+    {{if .Metadata.PackageName}}packageName: {{.Metadata.PackageName}}{{end}}
+    {{if .Metadata.GroupName}}groupName: {{.Metadata.GroupName}}{{end}}
+    {{if .Metadata.DepName}}depName: {{.Metadata.DepName}}{{end}}
+  description_body: ""
 ```
 
 ### Environment variable
