+++ pkg/controller/title.go
@@ -7,8 +7,6 @@ import (
 	"github.com/suzuki-shunsuke/renovate-issue-action/pkg/template"
 )
 
-const defaultIssueTitleTemplate = `Renovate Automerge Failure({{.RepoOwner}}/{{.RepoName}}): {{if .Metadata.GroupName}}{{.Metadata.GroupName}}{{else}}{{.Metadata.PackageName}}{{.Metadata.DepName}}{{end}} {{if .Metadata.PackageFileDir}}({{.Metadata.PackageFileDir}}){{end}}`
-
 type TitleParam struct {
 	RepoOwner string
 	RepoName  string
@@ -16,7 +14,7 @@ type TitleParam struct {
 }
 
 func getIssueTitle(cfg *config.Config, repoOwner, repoName string, metadata *Metadata) (string, error) {
-	s, err := template.Render(defaultIssueTitleTemplate, &TitleParam{
+	s, err := template.Render(cfg.Issue.Title, &TitleParam{
 		RepoOwner: repoOwner,
 		RepoName:  repoName,
 		Metadata:  metadata,
